backend/requirements.txt
Flask>=2.0.0
flask-cors>=4.0.0
spacy>=3.0.0,<3.9
stopwordsiso>=0.6.1
langdetect>=1.0.9
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
pytesseract>=0.3.10
docx2txt>=0.8
python-docx>=0.8.10
rarfile>=4.0
beautifulsoup4>=4.11.0
Pillow>=9.0.0
nltk>=3.7
matplotlib>=3.0.0
pandas>=1.0.0
SQLAlchemy>=2.0.0
Werkzeug>=2.0.0
